move batch-submitter types to core-utils
