+++ packages/batch-submitter/src/index.ts
@@ -1,3 +1,2 @@
 export * from './batch-submitter'
 export * from './transaction-chain-contract'
-export * from './types'
