+++ packages/core-utils/src/index.ts
@@ -4,3 +4,4 @@ export * from './watcher'
 export * from './base-service'
 export * from './l2context'
 export * from './events'
+export * from './batches'
